add internal catalyst-toolbox AdvisorReview type
