+++ src/catalyst-toolbox/catalyst-toolbox/src/types/mod.rs
@@ -1,2 +1,3 @@
 pub mod challenge;
 pub mod proposal;
+pub mod advisor_review;
